pyproject.toml
[project]
name = "rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gym-super-mario-bros>=7.4.0",
    "gymnasium[classic-control]>=0.29.1",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "pufferlib==2.0.6",
    "pyaml>=25.7.0",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
